Removed explicit casting by the QGauss instances' constructors.

+++ source/heat_equation/assemble_advection_matrix.cc
@@ -19,7 +19,7 @@ void HeatEquation<dim>::assemble_advection_matrix()
   // Polynomial degree of the integrand
   const int p_degree = velocity->fe_degree + 2 * temperature.fe_degree - 1;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
