+++ source/heat_equation/assemble_constant_matrices.cc
@@ -17,7 +17,7 @@ void HeatEquation<dim>::assemble_constant_matrices()
   // Polynomial degree of the integrand
   const int p_degree = 2 * temperature.fe_degree;
 
-  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * (p_degree + 1)));
+  const QGauss<dim>   quadrature_formula(std::ceil(0.5 * double(p_degree + 1)));
 
   using CellFilter =
     FilteredIterator<typename DoFHandler<dim>::active_cell_iterator>;
